requirements.txt
django-rest-framework
django
paystack
pymysql
firebase-admin
Pillow
corsheaders
```